Added identifier for specifying signin form redirect
This allows us to manage whether or not the page is redirected to the sign in form on a per-page basis with a variable, as opposed to a list of pages in the header.

Furthermore, and equally importantly, this allows us to change a page's preference for redirection on the same page that we're already modifying, without having to needlessly change a separate file.